Restructured test files
They now live alongside the file under test.
